change foreign key handler to use AR methods instead of SQL (#3)
* change foreign key handler to use AR methods instead of SQL
* added mysql
* added mariadb
* fix up postgresql versions

+++ spec/support/schema.rb
@@ -11,7 +11,7 @@ ActiveRecord::Schema.define do
 
   create_table :movies, force: :cascade do |t|
     t.string :name
-    t.integer :genre_id
+    t.bigint :genre_id
   end
 
   create_table :parent_models, force: :cascade do |t|
@@ -24,19 +24,19 @@ ActiveRecord::Schema.define do
 
   create_table :sub_parent_models, force: :cascade do |t|
     t.string :name
-    t.integer :parent_model_id
+    t.bigint :parent_model_id
   end
 
   create_table :simple_child_models, force: :cascade do |t|
     t.string :name
-    t.integer :parent_model_id
-    t.integer :second_parent_model_id
-    t.integer :sub_parent_model_id
+    t.bigint :parent_model_id
+    t.bigint :second_parent_model_id
+    t.bigint :sub_parent_model_id
   end
 
   create_table :polymorphic_child_models, force: :cascade do |t|
     t.string :name
-    t.integer :parent_id
+    t.bigint :parent_id
     t.string :parent_type
   end
 
@@ -45,7 +45,7 @@ ActiveRecord::Schema.define do
   end
 
   create_table :parent_models_simple_child_with_join_table_models, force: :cascade do |t|
-    t.integer :parent_model_id
-    t.integer :simple_child_with_join_table_model_id
+    t.bigint :parent_model_id
+    t.bigint :simple_child_with_join_table_model_id
   end
 end
